+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 [@Christopher-Hayes](github.com/Christopher-Hayes) - [@danyalaytekin](github.com/danyalaytekin) - [@flutterrausch](github.com/flutterrausch) - [@hakula139](github.com/hakula139) - [@lvii](github.com/lvii) - [@moritz-t-w](github.com/moritz-t-w) - [@nickv2002](github.com/nickv2002) - [@nossebro](github.com/nossebro) - [@PeterDaveHello](github.com/PeterDaveHello) - [@raphael2692](github.com/raphael2692) - [@rambalachandran](github.com/rambalachandran) - [@xmjiao](github.com/xmjiao) - [@wojtekcz](github.com/wojtekcz) - [@ZsgsDesign](github.com/ZsgsDesign) - [@zzy-life](github.com/zzy-life)
 
+## Next [next]
+
+- ✨ **Feature** - Added support for `gpt-4.1` when using OpenAI.
+- 📼 **Deprecated** - `gpt-4` is deprecated in favor of newer models.
+
 ## October 20, 2024 [v3.26.0]
 
 - ✨ **Feature** - Added support for `o1-preview` and `o1-mini` when using OpenAI.
@@ -142,7 +147,7 @@
 - 🎮 **QoL** - When opening code in a new text editor, VSCode should now automatically know how to syntax highlight it.
 - 🎮 **QoL** - Configuring "System message" is now an extension setting.
 
-[Not yet released]: https://github.com/vscode-reborn-ai/vscode-reborn-ai/compare/3.26.0...HEAD
+[next]: https://github.com/vscode-reborn-ai/vscode-reborn-ai/compare/3.26.0...HEAD
 [v3.26.0]: https://github.com/vscode-reborn-ai/vscode-reborn-ai/compare/3.25.0...3.26.0
 [v3.25.0]: https://github.com/vscode-reborn-ai/vscode-reborn-ai/compare/3.24.0...3.25.0
 [v3.24.0]: https://github.com/vscode-reborn-ai/vscode-reborn-ai/compare/3.23.2...3.24.0